more srp refactoring
