+++ freesas/app/auto_gpa.py
@@ -34,7 +34,7 @@ import logging
 from argparse import Namespace
 from freesas.autorg import auto_gpa
 from .sas_argparser import GuinierParser
-from .guinier_fitting import run_guinier_fit
+from .fitting import run_guinier_fit
 
 logging.basicConfig(level=logging.WARNING)
 logger = logging.getLogger("auto_gpa")
